rename api/fetcher to api/base and create the "posting" base function for making api post requests
