+++ utils/api/user.js
@@ -1,5 +1,5 @@
 import useSWR from 'swr'
-import { fetcher } from './fetcher'
+import { fetcher } from './base'
 
 // NOTE(alishaevn): this path is likely to change as a of result of next week's tech meeting, but this sets the foundation
 export const getCurrentUser = () => {
